requirements.txt
numpy
scipy
pyrtlsdr
